% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ONS_geolist.R
\name{ONS_geolist}
\alias{ONS_geolist}
\title{List ONS area code lists}
\usage{
ONS_geolist(area_type, date = NULL)
}
\arguments{
\item{area_type}{Type of area (eg. "Local Authority", "Region")}

\item{date}{The most recent date you would like a list for. This will return only those files which pre-date the date parameter.
If left blank, the function will return all valid file dates. This should be in the format eg."31-Dec-19"}
}
\value{
A vector of dates
}
\description{
This function generates a list of dates that relate to ONS area code lists.
ONS regularly update lists of geographical area codes. This allows you to see the date of each updated list for a particular area type.
For the area type given in 'area_type', the function will return a list of dates which relate to each updated list of codes in the ONS geographical data portal.
You should select the date that relates to your data and use that date in the geocheck() function call.
}
